back/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic-settings
httpx
python-dotenv
